ref(react): Refactor select text into utils
