+++ src/sentry/static/sentry/app/components/autoSelectText.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import jQuery from 'jquery';
 import ReactDOM from 'react-dom';
+import jQuery from 'jquery';
+
+import {selectText} from '../utils/selectText';
 
 const AutoSelectText = React.createClass({
   componentDidMount() {
@@ -15,23 +17,15 @@ const AutoSelectText = React.createClass({
 
   selectText() {
     let node = ReactDOM.findDOMNode(this.refs.element).firstChild;
-    if (document.selection) {
-      let range = document.body.createTextRange();
-      range.moveToElementText(node);
-      range.select();
-    } else if (window.getSelection) {
-      let range = document.createRange();
-      range.selectNode(node);
-      let selection = window.getSelection();
-      selection.removeAllRanges();
-      selection.addRange(range);
-    }
+    selectText(node);
   },
 
   render() {
+    let {className, children, style} = this.props;
+
     return (
-      <div ref="element" className={this.props.className}>
-        {this.props.children}
+      <div ref="element" className={className} style={style}>
+        {children}
       </div>
     );
   }
